requirements.txt
Django
matplotlib
psycopg2-binary
django-crispy-forms==1.9.2
